Show the smallest possible drawer parts
Instead of using squares that are as big as the whole doll, only
use the smallest square that fits all the parts in a given drawer.
That way, they look much more compact and nicer.

Also, as they are now drawn separately, they are shown as different
buttons, instead of a single, big image. Apart from looking better
(as it's now easy to style them separately) they can now flow in
the screen and adapt to different screen sizes.